Complete Query module Phase 2 refactoring with modular callback architecture
## Summary
- Successfully extracted and modularized query callbacks into organized structure
- Implemented ui/, callbacks/, helpers/ subdirectory architecture
- Fixed duplicate callback registration conflicts
- Enhanced data formatting and validation functions

## Architecture Improvements
- **Modular Callbacks**: Extracted 8 major callbacks from monolithic query.py (2,429 lines) into organized modules:
  - `query/callbacks/filters.py` - Age slider, demographic filters, live participant count
  - `query/callbacks/data_loading.py` - Data status, merge strategy, table options
  - `query/callbacks/state.py` - Store management and state updates
  - `query/callbacks/export.py` - Data export and download functionality (structure)

- **UI Components**: Organized reusable UI components in `query/ui/` with consistent styling and layout management

- **Helper Functions**: Centralized data formatting and validation in `query/helpers/` for better testability

## Key Fixes
- Resolved duplicate callback output conflicts by removing extracted callbacks from original pages/query.py
- Fixed query function call errors in live participant count functionality
- Enhanced phenotypic to behavioral filter conversion with proper error handling
- Improved callback registration system with centralized management

## Implementation Status
- ✅ Phase 1: Foundation and structure
- ✅ Phase 2: Callback decomposition (core callbacks extracted)
- 🔄 Phase 3-6: Remaining large callbacks and optimizations (planned)
