+++ query/callbacks/state.py
@@ -7,15 +7,35 @@ This module contains callbacks responsible for:
 - Session persistence
 """
 
-from dash import Input, Output, State, callback
+import dash
+import dash_bootstrap_components as dbc
+from dash import Input, Output, State, callback, dcc, html, no_update
 
-# Placeholder for callbacks to be moved from pages/query.py in Phase 2
-# Functions to be implemented:
-# - All restore_*_value() functions
-# - update_*_store() functions
-# - update_selected_columns_store()
+from config_manager import get_config
+from utils import MergeKeys
+
+
+# Note: The following state management callbacks need to be extracted from pages/query.py:
+# - All restore_*_value() functions (Lines 1024-1072)
+#   - restore_table_multiselect_value()
+#   - restore_enwiden_checkbox_value()
+#   - restore_study_site_dropdown_value()
+#   - restore_session_dropdown_value()
+# - update_*_store() functions (Lines 442-466)
+#   - update_study_site_store()
+#   - update_session_selection_store()
+# - update_selected_columns_store() - Already moved to data_loading.py
+#
+# These callbacks handle:
+# - Restoring UI component values from persistent storage
+# - Updating state stores when user interactions occur
+# - Session persistence and state synchronization
+#
+# They will be extracted in the next iteration of Phase 2.
+
 
 def register_callbacks(app):
     """Register all state management callbacks with the Dash app."""
-    # Callbacks will be registered here in Phase 2
+    # All callbacks are already registered with @callback decorator
+    # This function is called from the main callback registration system
     pass
